+++ doc/tutorial/vector_scal_cuda.cu
@@ -16,7 +16,7 @@
 
 #include <starpu.h>
 
-static __global__ void vector_mult_cuda(float *val, unsigned int n, float factor)
+static __global__ void vector_mult_cuda(float *val, size_t n, float factor)
 {
         unsigned i =  blockIdx.x*blockDim.x + threadIdx.x;
         if (i < n)
@@ -29,7 +29,7 @@ extern "C" void vector_scal_cuda(void *buffers[], void *cl_arg)
 	starpu_codelet_unpack_args(cl_arg, &factor);
 
         /* length of the vector */
-        unsigned int n = STARPU_VECTOR_GET_NX(buffers[0]);
+        size_t n = STARPU_VECTOR_GET_NX(buffers[0]);
         /* local copy of the vector pointer */
         float *val = (float *)STARPU_VECTOR_GET_PTR(buffers[0]);
         unsigned threads_per_block = 64;
