+++ examples/filters/custom_mf/custom_interface.c
@@ -76,7 +76,7 @@ static void     free_custom_buffer_on_node(void *data_interface, unsigned node);
 static size_t   custom_interface_get_size(starpu_data_handle_t handle);
 static uint32_t footprint_custom_interface_crc32(starpu_data_handle_t handle);
 static void     display_custom_interface(starpu_data_handle_t handle, FILE *f);
-static uint32_t custom_get_nx(starpu_data_handle_t handle);
+static size_t custom_get_nx(starpu_data_handle_t handle);
 
 
 static struct starpu_multiformat_data_interface_ops*get_mf_ops(void *data_interface)
@@ -245,10 +245,10 @@ static void display_custom_interface(starpu_data_handle_t handle, FILE *f)
 {
 	struct custom_data_interface *ci = (struct custom_data_interface *)
 		starpu_data_get_interface_on_node(handle, STARPU_MAIN_RAM);
-	fprintf(f, "Custom interface of size %u", ci->nx);
+	fprintf(f, "Custom interface of size %zu", ci->nx);
 }
 
-static uint32_t
+static size_t
 custom_get_nx(starpu_data_handle_t handle)
 {
 	struct custom_data_interface *data_interface;
@@ -261,7 +261,7 @@ custom_get_nx(starpu_data_handle_t handle)
 void custom_data_register(starpu_data_handle_t *handle,
 				 int home_node,
 				 void *ptr,
-				 uint32_t nx,
+				 size_t nx,
 				 struct starpu_multiformat_data_interface_ops *format_ops)
 {
 	struct custom_data_interface custom =
